value estimation: fix
Calculation of the first value of each interval was fixed.
It was necessary to compute the difference between the first
values of current and following interval.

Number of values in an estimated graph was edited
because it was necessary to take into consideration
the number of values after splitting them according to
a given interval.
Note: It is possible that some values were cut.

Signed-off-by: Klara <klarksonn@gmail.com>